Add a device test suite
A rather large commit, copied from a similar (almost identical suite) in
libtouchpad) and ported for libinput.

The goal here is to make testing for various devices easy, so the litest
("libinput test") wrappers do that.
check supports suites and test cases, both named. We wrap that so that each
named set of cases we add are a test suite, with the set of devices being the
test cases. i.e.

litest_add("foo:bar", some_test_function, LITEST_TRACKPOINT|LITEST_SYNAPTICS_CLICKPAD);

adds a suite named "foo:bar" and test cases for both devices given, with their
shortnames as test case name, resulting in:
   "foo:bar", "trackpoint"
   "foo:bar", "clickpad"

Multiple test functions can be added to a suite.

The environment variables CK_RUN_SUITE and CK_RUN_CASE can be used to narrow
the set of test cases. The test suite detects when run inside a debugger and
disables fork mode (the default).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>