Replace libinput_destroy() with libinput_unref()
Allow refcounting on the libinput context itself.

The main difference for refcounting here vs seat/event/device refcounting is
that the libinput context is still considered a top-level object. Once
destroyed, it will destroy all dependent objects. Thus, having a ref on an
event will not keep the libinput context alive once it is unref'd.

This function is needed for xorg drivers where we have multiple otherwise
independent devices share a context. To make sure this context can be
safely free()'d on device removal, each device ups the refcount and unrefs it
when removed.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>