path: modify backend to allow for more than one device
This brings the path backend closer in line with the udev backend. This is
needed for Xorg input drivers that need to share a libinput context between
multiple device instances but still make use of the server's device hotplug
mechanisms.

Sharing an instance is needed for future features like disabling the touchpad
while typing, etc.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>