udev: add libinput_udev_rescan_devices()
When a libinput context for a given seat is initialized, not all devices may
be avialable. Some or all devices may be paused by systemd-logind. Waiting for
a unplug event is not appropriate here as the devices are physically
available, just prevented from getting access.

Add a libinput_udev_rescan_devices() function that triggers a scan of all
devices on the current udev seat. Devices that do not exist on the seat are
added, devices that already exist on the seat but have been revoked are
removed.

Note that devices that have been physically removed are not removed, instead
we wait for the udev event to handle this for us.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>