touchpad: detect coordinate jumps on touchpads
Such jumps are usually consequences of the touchpad firmware failing to
notice that a different finger is in fact touching the touchpad.
This happens frequently when a finger is lifted at the same time as a
second finger is set down. Instead of a ABS_MT_TRACKING_ID  of -1
followed by a new touchpoing we get a large movement from the previous
position to the next position. This leads to a cursor jump on the
screen.

This patch adds detection for these jumps. Since the new finger is a new
touchpoint, all the usual rules for a new touch point should apply
(softbutton detection, tapping, etc.).

Thus, if we find one or more touchpoints to have jumped, we
mark the touchpoint as jumped and end it. Then we process the current
state, releasing all buttons where appropriate.
Then we re-process the state, marking all jumped and ended touchpoints
as TOUCH_BEGIN (and pressing buttons where appropriate).

The jump threshold is 20mm on devices with resolution, or a
quarter of the diagonal for devices without a resolution.

https://bugs.freedesktop.org/show_bug.cgi?id=76722

Based on a patch by Alexander E. Patrakov <patrakov@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>