touchpad: Add tp_button_touch_active function
We don't want touches in the button area to cause the pointer to move, add
a tp_button_touch_active function which the main code in evdev-mt-touchpad
can call to see if a touch should be consider a candidate for being the
pointer, should be taken into account for 2 finger scrolling, etc.

The idea behind the main code polling for this is that in the future with
ie edge scrolling we will have another independent state machine, which
may also want to block a touch from being the pointer, so it is best for
the main code to test all independent state machines, rather then having
the state-machines poke the is_pointer variabel directly.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>