touchpad: Rework is_pointer handling
Move scanning for a suitable touch to be the pointer to tp_process_state
and take tp_button_touch_active into account.

Note this adds a tp_touch_active helper since we want to do the same checks in
other places too (e.g. to see if a finger should count for 2 finger scrolling).

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>

Squashed in: if 0 hunk

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>