Restore capability events
This reverts commit ab9260c5c70b95779f39069b31646ff5cf8970ad, with a bunch of
changes on top.

First change is different naming, the events are now DEVICE_CAPABILITY_ADDED
and DEVICE_CAPABILITY_REMOVED for consistency with the device events.

The behaviour of capability events is as previously, the event sequence is
always: DEVICE_ADDED
	DEVICE_CAPABILITY_ADDED
	DEVICE_CAPABILITY_ADDED
	...
	DEVICE_CAPABILITY_REMOVED
	DEVICE_CAPABILITY_REMOVED
	DEVICE_REMOVED

Disabling a device does not send capablity events, but suspending a libinput
context does (since it removes the device).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>