touchpad: re-route trackpoint buttons on the *50 Lenovo series
The laptops on this series have the physical trackpoint buttons back but
wired them up to the touchpad instead of the trackpoint device and they appear
as BTN_0, BTN_1 and BTN_2 for left, right, middle.

The udev hwdb marks these for us with the TOUCHPAD_TRACKPOINT_BUTTONS tag. Use
that tag to identify them and re-route the events through the trackstick
device after mangling the event codes to represent the actual buttons.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>