tablet: ignore tools already in proximity at startup
Require a tool to be moved into proximity before we handle any events from it.
This requires the user to lift the tool for first use but a tool that's
already on the tablet on init is likely to send garbage anyway and interfere.
This way users can leave the mouse/pen on the tablet overnight and their
events won't interfere until they actually use it.

Reported-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>