Push the touchpad magic slowdown to the touchpad accel code
This way the unaccelerated deltas returned by libinput are correct.

To maintain the current behavior we slow down the input speed by the magic
factor and likewise the accelerated output speed. This produces virtually the
same accelerated deltas as the previous code.

On touchpads without resolution, this will slow down the pointer movement.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>