touchpad: add pressure-based thumb-detection
All touchpad recordings seen so far show that a value above 100 is definitely
a thumb or a palm. Values below are harder to discern, and the same isn't true
for touchpads supporting ABS_PRESSURE instead of ABS_MT_PRESSURE.

Note that "thumb" is the synonym for "this touch is big enough to not be
a normal finger" which obviously means that a thumb-tip will still count as
normal finger.

The handling of a touch is as outlined in tp_thumb_detect:
* thumbs are ignored for pointer motion
* thumbs cancel gestures
* thumbs are ignored for clickfinger count
* edge scrolling doesn't care either way
* software buttons don't care either way
* tap: only if thumb on begin

The tap state machine could be update with a bunch of escape routes when a
thumb is detected in the middle of the state machine but this gets complex
quick. Let's see if there's a need for it first.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>