pad: add support for a get_mode() function on tablet pad events
Move mode control to libinput. This reduces some flexibility on what we can do
with modes but makes it a lot easier for anyone to implement modes correctly
and have the LEDs apply appropriately, etc. Let's go with the option to make
the 95% use-case easy.

Each tablet pad event has a _get_mode() call that returns the numeric
0-indexed mode for the source of this event, i.e. the button,ring, or strip.
Note: whether the mode is actually used is up to the caller, e.g. under
Windows and OS X the mode only applies to the rings/strips, not the buttons.

libinput will switch the LEDs accordingly.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>