pad: implement mode-switching for tablet pads
The Wacom pad leds are in a syspath in the wacom hid device. At least two
toggles per pad status_led0_select (which LED is on) and status0_luminance
(for brightness), some tablets have the led1 equivalent too.
Mode-switching requires libwacom for us to know how many modes to initialize.

We init a pad_led_group struct for each group and a pad_mode_button struct for
all buttons/rings/strips. The mode buttons have a target mode which is either
their numeric mode (24HD) or just NEXT (all others). Most buttons have a NONE
target mode, i.e. they don't toggle the mode on press. This is a bit over the
top, some more targeted approach than running through a list on button presses
could work too but for now this will do.

This does not work yet for the EKR which has the leds stored in a different
sysfs file.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>