test: remove NOTPARALLEL
We can now run the tests in parallel with make check -j <n>, even though results are
mixed: on an i7 here make -j 4 is reliable, on a under-resourced VM running on
an i5 make -j2 is the best I can do, anything higher fails with timeouts
possibly because of the udev lock contention (see 030ec053fbbc).

There is one test in device.test (device_reenable_syspath_changed) that fails
if another device is added/removed at the wrong time and it's reliable to
trigger that failure. The tests are split up into parallel and serial tests,
for the serial tests we have to manually add dependencies to avoid executing
them in parallel with others.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>