test: only init the device rules once per binary
The udev hwdb takes about 200ms and we still trigger it on each device. Plus
it requires us to ensure locking is correct across multiple binaries when
running tests in parallel. The udev rules don't actually change after
compiling, so simply create them once per binary. Rules are namespaced with the
binary name again, the practical effect is that if we do run tests in parallel
we'll have a bunch of identical rules file.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>