test: change the udev rules to be per-binary
We always install the same udev rule anyway but to avoid having to synchronise
udev rule installation and removal between independently started processes
just store each rule with the binary as suffix. So device.test will create a
rule with "blah-device.test.rule", etc.

And since we're now messing with filenames a lot more, store them in a list
and simply unlink them at the end.

This only affects the global udev rules, the ones per-device are not
namespaced yet.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>