test: fork off tests by suite group name
Run through the test suites which all have a group name in the form
"touchpad:motion". Use the bit before the colon as group name ("touchpad" in
this example) and fork off a subprocess to only run those tests.
Note that this uses the primitive approach, so if the group order in the list
is A A A B B B A A B B we end up with two processes for A and B each.

Each subprocess overwrites argv[0] to be easier identifiable in the ps
output when we're trying to figure out which tests are still running.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>