evdev: change hysteresis to work across both axes
The previous approach causes axis snapping, preventing things like drawing a
circle, see e.g. https://bugs.freedesktop.org/show_bug.cgi?id=99410

Switch to an approach that honors both axes. We know that whenever we're
outside the hysteresis margin, we move the center back by exactly "margin" for
the axis outside the margin. That's easy for a single axis, but for two axes
movements, we need to also calculate the amount of the other axis to move the
center back (less than margin).

This amount can be expressed as a ratio of the margin m and the x/y axis of
the movement vector:
	offset = m/(y/x)
Plus the extra bits to handle the signs and x > y vs. y < x, etc.

In the simplest case, a 45 degree movement will move back by (-m, -m). A
22.5 degree movement will move back by (-m, -m/2) and so on.

There are potential issues:
a movement of x/y where x > margin but y < margin now still moves along the y
axis, effectively ignoring the hysteresis here and potentially wobbling the
cursor along the y axis. Without that behavior we would get axis snapping
though (the issuue this patch is supposed to fix).

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>