Add an evdev fd interface to read events directly
This interface is not to be used for normal applications, its sole purpose is
to make timing-sensitive tests possible. After a device has been initialized
it swaps the real fd over for the provided event fd. libinput uses
event timestamps, so we can provide reliable event streams not possible when
going through uinput.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>